requirements.txt
pandas==2.3.3
plotly==6.5.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
SQLAlchemy==2.0.45
streamlit==1.53.0